refactor: Race에서 관리하는 시도횟수를 TryTime라는 일급컬렉션으로 리팩토링
race에서 시도횟수가 초기화 되었는데 다시 초기화 하려 할 경우, 에러 반환하는 기능 추가
